Routing and split css pages
