+++ src/components/Inicio.js
@@ -1,0 +1,13 @@
+import React from "react";
+import '../style/inicio.css';
+
+const Inicio = () => {
+    return (
+        <>
+        <div className="container1">
+        <h1 className="display-1 titulo">reelKeeper</h1>
+        </div>
+        </>
+    )
+}
+export default Inicio;
