+++ src/style/App.css
@@ -2,30 +2,33 @@
   @import url('https://fonts.googleapis.com/css2?family=JetBrains+Mono&display=swap');
 
 body{
-  /* background: rgb(240,240,240); */
-  background: linear-gradient(311deg, #f6f7f5, #aefff5, #ffb7ec, #fdffa0, #f99b78, #b862d2);
+  background: rgb(217, 217, 217);
+  /* background: linear-gradient(311deg, #f6f7f5, #aefff5, #ffb7ec, #fdffa0, #f99b78, #b862d2);
     background-size: 1200% 1200%;
 
     -webkit-animation: AnimationName 22s ease infinite;
     -moz-animation: AnimationName 22s ease infinite;
-    animation: AnimationName 22s ease infinite;
+    animation: AnimationName 22s ease infinite; */
+}
+.btn{
+  position: absolute;
+  top:auto;
+  left: 20%;
+  right: 40%;
+  bottom: 40%;
+  width: fit-content;
 }
 
-@-webkit-keyframes AnimationName {
-  0%{background-position:0% 54%}
-  50%{background-position:100% 47%}
-  100%{background-position:0% 54%}
+.loginButton{
+  position: absolute;
+  top:auto;
+  left: 50%;
+  right: 10%;
+  bottom: 40%;
+  width: fit-content;
+  background-color: #522377 ;
 }
-@-moz-keyframes AnimationName {
-  0%{background-position:0% 54%}
-  50%{background-position:100% 47%}
-  100%{background-position:0% 54%}
-}
-@keyframes AnimationName {
-  0%{background-position:0% 54%}
-  50%{background-position:100% 47%}
-  100%{background-position:0% 54%}
-}
+
 
 .movie-app > .row {
   /* overflow-x: auto; */
@@ -80,49 +83,9 @@ a{
   color: white;
   text-decoration: none;
 }
-/* HEADER */
 
-nav{
-  background-color: rgba(240, 240, 240, 0.546);
-}
-.navbar{
-  --bs-navbar-padding-y: 0rem;
-  --bs-navbar-padding: 0rem;
   
-  }
-
-  /* .fondoNav{
-    background-color: rgb(223, 223, 223);
-  } */
-  .borde{
-      width: 100px;
-      height: auto;
-      text-align: center;
-
-  }
-  .purple{
-      background-color: rgb(82,35,119);
-  }
-  .azul{
-      background-color: rgb(86,155,206);
-  }
-  .verde{
-      background-color: rgb(86,206,141);
-  }
-  .amarillo{
-      background-color: rgb(203,206,86);
-  }
-  .rojito{
-      background-color: rgb(206,144,86);
-  }
-  .blanco{
-      background-color: rgb(240,240,240);
-  }
-
-  /* FIN HEADER */
-
-
-  /* SCROLLBAR  */
+  /******************** SCROLLBAR  ***********************/
   /* width */
 ::-webkit-scrollbar {
   width: 10px;
